refactor: move SEO utility functions to a separate file and clean up console logging in fetch_images script
