+++ src/App.tsx
@@ -10,7 +10,8 @@ import Reservation from './components/Reservation';
 import Contact from './components/Contact';
 import Footer from './components/Footer';
 import { useMotionPreference } from './hooks/useMotionPreference';
-import Seo, { buildHomeJsonLd, buildOgImage } from './components/SEO';
+import Seo from './components/SEO';
+import { buildHomeJsonLd, buildOgImage } from './components/seoUtils';
 import Legal from './components/legal/Legal';
 import Privacy from './components/legal/Privacy';
 import Terms from './components/legal/Terms';
